Fix nil pointer dereference when an invalid file is requested (bsc#1199257)
The apiserver can dereference nil pointer when an invalid file is
requested but its os.Stat() check fails with a different error than
ErrNotExist. For instance, this can be reproduced by passing NUL in the
address, such as 'wget https://<hawk-ip>:7630/%00/'.

Correct the check in (*routeHandler).serveFile() to address this
problem.